Refactored the class and fixed phpcs errors
